Composer: the attachments + presents a system menu
The + opened a hand-rolled card: a coordinator relayed the tap up to
ConversationView, which floated the card from its root overlay against
the composer's published bounds, over a scrim. All of that existed
because the bar lives in a safeAreaBar that would clip any overlay of
the composer's own -- and it bought a menu with none of the system's
feel: a 0.92-scale fade instead of the menu spring, no haptic, no
dimming, no drag-to-select.

A system Menu presents in its own window, so the clipping constraint
never applies and the chrome problem that ruled out .popover (its
arrow) doesn't exist. The + now presents one directly: same rows, same
greyed-not-dropped disabled states, same debug-injector gating, and the
coordinator, the environment key, and the hosted card all go away. The
participation bubble keeps its card, which now owns ComposerMenu's
metrics alone.
